# src/apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Moviedle project",
        "title": "Moviedle",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/movies/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Allow a user to see details of a movie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Movie"
                ],
                "summary": "View movie details.",
                "operationId": "Movies.Get",
                "parameters": [
                    {
                        "type": "string",
                        "default": "a92adbf5-930c-41ce-bff3-b11798505f1c",
                        "description": "Movie ID.",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success.",
                        "schema": {
                            "$ref": "#/definitions/response.Movie"
                        }
                    },
                    "400": {
                        "description": "Bad Request.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal error.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    },
                    "503": {
                        "description": "Database out of function.",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "response.Director": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.ErrorMessage": {
            "type": "object",
            "properties": {
                "invalid_fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.InvalidField"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
        "response.Genre": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "response.InvalidField": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "field_name": {
                    "type": "string"
                }
            }
        },
        "response.Movie": {
            "type": "object",
            "properties": {
                "cover_url": {
                    "type": "string"
                },
                "directors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Director"
                    }
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Genre"
                    }
                },
                "id": {
                    "type": "string"
                },
                "oscars": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Oscar"
                    }
                },
                "resume": {
                    "type": "string"
                },
                "slogan": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "response.Oscar": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}